Renamed View3D to BaseView3D, and moved BaseView & BaseView3D into /views
