+++ source/views/BaseView.cpp
