+++ source/views/IwatodaiStreetsView.cpp
@@ -87,7 +87,7 @@ void IwatodaiStreetsView::setupEnvironment()
 
 void IwatodaiStreetsView::init()
 {
-    View3D::init();
+    BaseView3D::init();
 
     // setup sub screen
     bgSharedSub1 = bgInitSub(0, BgType_Text8bpp, BgSize_T_256x256, 2, 1);
